requirements.txt
streamlit==1.33.0
pandas==2.2.0
scikit-learn==1.4.0
sentence-transformers==2.5.0
google-generativeai==0.3.1